+++ plugins/groovy-wakamiti-plugin/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
 
+## [2.9.0] - Unreleased
+
+### Added
+- New configuration property `groovy.compilation.failOnError` (default: `true`) to abort test
+  execution when Groovy source files cannot be compiled. Set it to `false` to log the errors and
+  continue loading the remaining sources.
+
+
 ## [2.8.0] - 2026-07-31
 
 ### Changed